Move commands to separate files in Commands folder
