+++ Commands/bot_info.py
@@ -1,0 +1,13 @@
+import time
+
+
+def reply_with_bot(self, message):
+
+    text = 'BluePAGMANBot is a bot made by @Bluepigman5000 in Python. \
+    It has some basic commands, and can run a game of chess in chat \
+    between two different players. It is currently running on a VM on Google Cloud Compute Engine.'
+
+    if (message.user not in self.state or time.time() - self.state[message.user] >
+            self.cooldown):
+        self.state[message.user] = time.time()
+        self.send_privmsg(message.channel, text)
