+++ bot.py
@@ -8,17 +8,14 @@ to play a game of chess against another chatter.
 import socket
 import sys
 import ssl
-import datetime
 from collections import namedtuple
 import time
 import random
 import config
 import chess
-import chessCommands
-from threading import Timer
-import newsCommands
 from pymongo.mongo_client import MongoClient
-from Commands import *
+from Commands import ( bot_info, date, ping, help_chess, source_code, play_chess, ro, r960, help_ro, pyramid, slow_pyramid,
+news, help_news, daily, roulette, balance, leaderboard, help)
 
 
 
@@ -58,29 +55,30 @@ class Bot:
         self.choseSidePlayer1 = False
         self.currentGame = None  # hold the chess game
         self.chessTimer = None  # chess game pending timer
+        self.start_time = None # to measure bot uptime
+
         # anyone can use these
         self.custom_commands = {
             'date': date.reply_with_date,
             'ping': ping.reply_to_ping,
-            'help': self.list_commands,
-            'help_chess': self.reply_with_chesshelp,
-            'source_code': self.reply_with_source_code,
-            'play_chess': self.play_chess,
-            'bot': self.reply_with_bot,
-            'ro': self.reply_with_random_opening,
-            'joke': self.reply_with_joke,
-            'r960': self.reply_with_random960,
-            'help_ro': self.reply_with_help_ro,
-            'pyramid': self.reply_with_pyramid,
-            'slow_pyramid': self.reply_with_slow_pyramid,
-            'news': self.reply_with_news,
-            'help_news': self.reply_with_help_news,
-            'daily': self.reply_with_daily,
-            'roulette': self.reply_with_roulette,
-            'balance': self.reply_with_balance,
-            'leaderboard': self.reply_with_leaderboard
+            'help_chess': help_chess.reply_with_chesshelp,
+            'source_code': source_code.reply_with_source_code,
+            'play_chess': play_chess.play_chess,
+            'bot': bot_info.reply_with_bot,
+            'ro': ro.reply_with_random_opening,
+            'r960': r960.reply_with_random960,
+            'help_ro': help_ro.reply_with_help_ro,
+            'pyramid': pyramid.reply_with_pyramid,
+            'slow_pyramid': slow_pyramid.reply_with_slow_pyramid,
+            'news': news.reply_with_news,
+            'help_news': help_news.reply_with_help_news,
+            'daily': daily.reply_with_daily,
+            'roulette': roulette.reply_with_roulette,
+            'balance': balance.reply_with_balance,
+            'leaderboard': leaderboard.reply_with_leaderboard,
+            'help': help.list_commands
         }
-
+        # self.custom_commands[help] = self.list_commands
         # only bot owner can use these commands
         self.private_commands = {
             'leave': self.leave,
@@ -127,6 +125,7 @@ class Bot:
         for channel in self.channels:
             self.send_command(f'JOIN #{channel}')
             self.send_privmsg(channel, 'forsenEnter')
+        self.start_time = time.time()
         self.loop_for_messages()
 
     def get_user_from_prefix(self, prefix):
@@ -221,7 +220,7 @@ class Bot:
 
             # Aliases.
             if message.text_command.lower() == "commands":
-                self.custom_commands["help"](message)
+                self.custom_commands["help"](self, message)
                 self.time = time.time()
 
     def loop_for_messages(self):
@@ -229,336 +228,6 @@ class Bot:
             received_msgs = self.irc.recv(4096).decode(errors='ignore')
             for received_msg in received_msgs.split('\r\n'):
                 self.handle_message(received_msg)
-
-    """ General Commands here"""
-
-    # def reply_with_date(self, message):
-    #     if (message.user not in self.state or time.time() - self.state[message.user] >
-    #             self.cooldown):
-    #         self.state[message.user] = time.time()
-    #         formatted_date = datetime.datetime.now().strftime('%Y/%m/%d %H:%M:%S')
-    #         text = f'{message.user}, the date is {formatted_date} EST.'
-    #         self.send_privmsg(message.channel, text)
-
-    # def reply_to_ping(self, message):
-    #     if (message.user not in self.state or time.time() - self.state[message.user] >
-    #             self.cooldown):
-    #         self.state[message.user] = time.time()
-    #         text = f'@{message.user}, forsenEnter'
-    #         self.send_privmsg(message.channel, text)
-
-    def reply_with_source_code(self, message):
-        if (message.user not in self.state or time.time() - self.state[message.user] >
-                self.cooldown):
-            self.state[message.user] = time.time()
-            text = 'Source code: https://github.com/BluePigman/BluePagmanBot'
-            self.send_privmsg(message.channel, text)
-
-    def list_commands(self, message):
-        custom_cmd_names = list(self.custom_commands.keys())
-        all_cmd_names = [
-            self.command_prefix + cmd
-            for cmd in custom_cmd_names
-        ]
-        text = "" f'@{message.user}, Commands: ' + ' '.join(all_cmd_names)
-        if (message.user not in self.state or time.time() - self.state[message.user] >
-                self.cooldown):
-            self.state[message.user] = time.time()
-            self.send_privmsg(message.channel, text)
-
-    def reply_with_bot(self, message):
-
-        text = 'BluePAGMANBot is a bot made by @Bluepigman5000 in Python. \
-        It has some basic commands, and can run a game of chess in chat \
-        between two different players. It is currently running on a VM on Google Cloud Compute Engine.'
-
-        if (message.user not in self.state or time.time() - self.state[message.user] >
-                self.cooldown):
-            self.state[message.user] = time.time()
-            self.send_privmsg(message.channel, text)
-
-    def reply_with_random_opening(self, message):
-
-        text = f'@{message.user}, '
-
-        if (message.user not in self.state or time.time() - self.state[message.user] >
-                self.cooldown):
-            self.state[message.user] = time.time()
-
-            # if args present
-            if (message.text_args):
-
-                if '-w' in message.text_args:
-                    # get opening for white
-                    side = 'w'
-                    message.text_args.remove('-w')
-                    name = " ".join(message.text_args)
-                    if '\U000e0000' in name:
-                        name = name.replace('\U000e0000', '')
-
-                    opening = chessCommands.getRandomOpeningSpecific(
-                        name, side)
-                    self.send_privmsg(message.channel, text + opening)
-
-                elif '-b' in message.text_args:
-                    # get opening for black
-                    side = 'b'
-                    message.text_args.remove('-b')
-                    name = " ".join(message.text_args)
-                    if '\U000e0000' in name:
-                        name = name.replace('\U000e0000', '')
-
-                    opening = chessCommands.getRandomOpeningSpecific(
-                        name, side)
-                    self.send_privmsg(message.channel, text + opening)
-
-                else:  # get opening for specified search term
-                    name = " ".join(message.text_args)
-                    if '\U000e0000' in name:
-                        name = name.replace('\U000e0000', '')
-
-                    opening = chessCommands.getRandomOpeningSpecific(name)
-                    self.send_privmsg(message.channel, text + opening)
-
-            else:  # No arguments
-                opening = chessCommands.getRandomOpening()
-                self.send_privmsg(message.channel, text + opening)
-
-    def reply_with_random960(self, message):
-        if (message.user not in self.state or time.time() - self.state[message.user] >
-                self.cooldown):
-            self.state[message.user] = time.time()
-            opening = chessCommands.getRandom960()
-            self.send_privmsg(message.channel, opening)
-
-    def reply_with_joke(self, message):
-        if (message.user not in self.state or time.time() - self.state[message.user] >
-                self.cooldown):
-            self.state[message.user] = time.time()
-            setup, punchline = chessCommands.getJoke(self.joke)
-            self.send_privmsg(message.channel, setup)
-            punchline = punchline.strip('\n')
-            x = random.randint(0, 10)
-            if x > 1:
-                response = f'{punchline} haHAA'
-            else:
-                response = f'{punchline} Pepepains'
-            time.sleep(3)
-            self.send_privmsg(message.channel, response)
-            self.joke += 1
-            if self.joke > 84:
-                self.joke = 0
-
-    # https://stackoverflow.com/q/6266727
-    def reply_with_pyramid(self, message):
-        if (message.user not in self.state or time.time() - self.state[message.user] >
-                self.cooldown):
-            self.state[message.user] = time.time()
-            if len(message.text_args) > 1 and message.text_args[1].isnumeric():
-                emote = message.text_args[0]
-                width = int(message.text_args[1])
-                if len(emote) * width + width - 1 < 500:
-                    text = ''
-                    for x in range(width):  # go up
-                        text += (emote + ' ')
-                        self.send_privmsg(message.channel, text)
-                        time.sleep(0.1)
-                    for y in range(width):  # go down
-                        text = text.rsplit(emote, 1)[0]
-                        self.send_privmsg(message.channel, text)
-                        time.sleep(0.1)
-                else:
-                    text = 'Pyramid is too large to be displayed in chat. Use \
-                    a smaller pyramid width.'
-                    self.send_privmsg(message.channel, text)
-
-            else:
-                text = f'Width must be an integer. Usage: {self.command_prefix}pyramid {{name}} {{width}}'
-                self.send_privmsg(message.channel, text)
-
-    def reply_with_slow_pyramid(self, message):
-        if (message.user not in self.state or time.time() - self.state[message.user] >
-                self.cooldown):
-            self.state[message.user] = time.time()
-            if len(message.text_args) > 1 and message.text_args[1].isnumeric():
-                emote = message.text_args[0]
-                width = int(message.text_args[1])
-                if len(emote) * width + width - 1 < 500:
-                    text = ''
-                    for x in range(width):  # go up
-                        text += (emote + ' ')
-                        self.send_privmsg(message.channel, text)
-                        time.sleep(1.6)
-                    for y in range(width):  # go down
-                        text = text.rsplit(emote, 1)[0]
-                        self.send_privmsg(message.channel, text)
-                        time.sleep(1.6)
-                else:
-                    text = 'Pyramid is too large to be displayed in chat. Use \
-                    a smaller pyramid width.'
-                    self.send_privmsg(message.channel, text)
-
-            else:
-                text = f'Width must be an integer. Usage: {self.command_prefix}pyramid {{name}} {{width}}'
-                self.send_privmsg(message.channel, text)
-
-
-    def reply_with_news(self, message):
-        if (message.user not in self.state or time.time() - self.state[message.user] >
-                self.cooldown):
-            self.state[message.user] = time.time()
-            try:
-                m = newsCommands.get_random_news_item()
-                if (message.text_args):
-                    keywords = ' '.join(message.text_args)
-                    if '\U000e0000' in keywords:
-                        keywords = keywords.replace('\U000e0000', '')
-                    keywords = keywords.replace(" ", "+")
-                    keywords = keywords.replace("#", "'#'")
-                    keywords = keywords.replace("&", "'&'")
-                    m = newsCommands.get_random_news_item(keywords)
-                self.send_privmsg(message.channel, m)
-            except Exception as e:
-                print(e)
-                self.send_privmsg(message.channel, f"@{message.user}, No news found for the given query.")
-
-    def reply_with_help_news(self, message):
-        if (message.user not in self.state or time.time() - self.state[message.user] >
-                self.cooldown):
-            self.state[message.user] = time.time()
-            self.send_privmsg(message.channel, newsCommands.get_help_text())
-
-    def reply_with_daily(self, message):
-        if (message.user not in self.state or time.time() - self.state[message.user] >
-                self.cooldown):
-            self.state[message.user] = time.time()
-            user_data = self.users.find_one({'user': message.user})
-            if not user_data: # add new user
-                self.users.insert_one({'user': message.user , 'last_claimed':datetime.datetime.now(), 'points': 100 })
-                self.send_privmsg(message.channel, f'@{message.user}, your daily reward is 100 Pigga Coins. Come back tomorrow for more!')
-            else: # check time
-                last_claimed = user_data['last_claimed']
-                time_diff = datetime.datetime.now() - last_claimed
-                if time_diff.days >= 1: # give reward
-                    self.users.update_one({'user': message.user}, {'$set': {'last_claimed': datetime.datetime.now()}, '$inc': {'points': 100}})
-                    self.send_privmsg(message.channel, f'@{message.user}, your daily reward is 100 Pigga Coins. Come back tomorrow for more!')
-                else:
-                    remaining_time = datetime.timedelta(days=1) - time_diff
-                    t = "You have already collected your dailies today. Come back in {}.".format(remaining_time)
-                    self.send_privmsg(message.channel, t)
-
-    def reply_with_roulette(self, message):
-        if (message.user not in self.state or time.time() - self.state[message.user] >
-                self.cooldown):
-            self.state[message.user] = time.time()
-            user_data = self.users.find_one({'user': message.user})
-
-            if not user_data: 
-                m = f"@{message.user}, you do not have any Pigga Coins. Use the daily command."
-                self.send_privmsg(message.channel, m)
-            else:
-                user_points = user_data['points']
-                amount = message.text_args[0]
-
-                if user_points == 0:
-                    m = f"@{message.user}, you don't have any Pigga Coins. PoroSad"
-                    self.send_privmsg(message.channel, m)
-                else:
-                    if amount == 'all':
-                        amount = user_points
-                    elif not amount.isdigit() or int(amount) <= 0:
-                        m = f"@{message.user}, please enter a positive number or 'all'."
-                        self.send_privmsg(message.channel, m)
-                        return  # Exit the function early if the input is invalid
-
-                    amount = int(amount)  # Convert the amount to an integer
-
-                    if amount > user_points:
-                        m = f"@{message.user}, you don't have enough Pigga Coins."
-                        self.send_privmsg(message.channel, m)
-                    else:
-                        gamba = random.randint(1, 2)
-                        if gamba == 1:
-                            self.users.update_one({'user': message.user}, {'$inc': {'points': amount}})
-                            new_balance = user_points + amount
-                            m = f'@{message.user}, you won {amount} Pigga Coins in roulette and now have {new_balance} Pigga Coins!'
-                            self.send_privmsg(message.channel, m)
-                        else:
-                            amountDec = -amount
-                            self.users.update_one({'user': message.user}, {'$inc': {'points': amountDec}})   
-                            new_balance = user_points - amount
-                            m =  f'@{message.user}, you lost {amount} Pigga Coins in roulette and now have {new_balance} Pigga Coins! Saj'
-                            self.send_privmsg(message.channel, m)
-
-
-    def reply_with_balance(self, message):
-        if (message.user not in self.state or time.time() - self.state[message.user] >
-                self.cooldown):
-            self.state[message.user] = time.time()
-            if message.text_args:
-                searchUser = message.text_args[0]
-                if '\U000e0000' in searchUser:
-                    searchUser = searchUser.replace('\U000e0000', '')
-                if '@' in searchUser:
-                    searchUser = searchUser.replace('@', '')
-                user_data = self.users.find_one({'user': searchUser.lower()})
-                if not user_data:
-                    self.send_privmsg(message.channel, f'@{message.user}, That user is not in the database.')
-                else:
-                    self.send_privmsg(message.channel, f'@{message.user}, {searchUser} has {user_data["points"]} Pigga Coins.')
-
-
-            else:
-                user_data = self.users.find_one({'user': message.user})
-                if not user_data:
-                    self.send_privmsg(message.channel, f'@{message.user}, you do not have any Pigga Coins. Use the daily command.')
-                else:
-                    self.send_privmsg(message.channel, f'@{message.user}, you have {user_data["points"]} Pigga Coins.')        
-
-    def reply_with_leaderboard(self, message):
-        top_users = self.users.find().sort([('points', -1)]).limit(5)
-        leaderboard_message = "Leaderboard: "
-        for idx, user in enumerate(top_users, start=1):
-            leaderboard_message += f"{idx}. @{user['user']} - {user['points']}   "
-        self.send_privmsg(message.channel, leaderboard_message)
-
-    """ Chess commands """
-
-    def reply_with_chesshelp(self, message):
-
-        text = (f"""@{message.user}, moves should be entered in either SAN or UCI notation. SAN moves \
-            follow the format: letter of piece (except for pawn moves), x if there is a capture, \
-            and the coordinate of the square the piece moves to. \
-            For pawn promotions add = followed by the letter of the piece. \
-            Examples: f6, Nxg5, <Ke2, a1=Q, bxc8=R, Bh8.""")
-
-        text2 = "Sometimes you need to indicate the exact \
-            piece that is moving if there is ambiguity. Examples include Nge2, Rhxe1. To castle, enter O-O or O-O-O. \
-            Refer to https://en.wikipedia.org/wiki/Portable_Game_Notation#Movetext for more detailed information. "
-
-        text3 = f"""UCI moves follow the format: original coordinate of piece, new coordinate of piece. \
-            For castling, use the king's coordinates. UCI Input must be in lowercase and contain no spaces.\
-            For example, if you want to start by moving the e pawn to e4, \
-            you type {self.command_prefix}move e4 OR {self.command_prefix}move e2e4 \
-            To resign type {self.command_prefix}move resign"""
-
-        if (message.user not in self.state or time.time() - self.state[message.user] >
-                self.cooldown):
-            self.state[message.user] = time.time()
-            self.send_privmsg(message.channel, text)
-            time.sleep(2)
-            self.send_privmsg(message.channel, text2)
-            time.sleep(2)
-            self.send_privmsg(message.channel, text3)
-
-    def reply_with_help_ro(self, message):
-        text = (f"@{message.user}, Gets a random opening. You can add -b or -w \
-for a specific side, and/or add a name for search. e.g. {self.command_prefix}ro King's Indian \
-                Defense -w")
-        if (message.user not in self.state or time.time() - self.state[message.user] >
-                self.cooldown):
-            self.state[message.user] = time.time()
-            self.send_privmsg(message.channel, text)
 
     """Private commands"""
 
@@ -636,16 +305,6 @@ for a specific side, and/or add a name for search. e.g. {self.command_prefix}ro 
             time.sleep(2)
         pass
 
-    def play_chess(self, message):  # start a game of chess
-
-        if not self.chessGameActive:
-            self.chessGameActive = True
-            self.player1 = message.user
-            text = f'@{self.player1} has started a chess game. Type {self.command_prefix}join to join \
-                the game.'
-            self.send_privmsg(message.channel, text)
-            self.chessTimer = Timer(30, self.gameTimeout, (message.channel,))
-            self.chessTimer.start()  # start a timer of 30s.
 
     def chooseSidePlayer1(self, message):
         # Player who started game chooses side first.
